refactor formatting of fixed-point numbers
